Further organize javarosa plugin code
