+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/JavaRosaInitializer.kt
@@ -10,8 +10,8 @@ import org.javarosa.xform.util.XFormUtils
 import org.odk.collect.android.dynamicpreload.DynamicPreloadXFormParserFactory
 import org.odk.collect.android.entities.EntitiesRepositoryProvider
 import org.odk.collect.android.logic.actions.setgeopoint.CollectSetGeopointActionHandler
-import org.odk.collect.entities.javarosa.EntityXFormParserFactory
-import org.odk.collect.entities.javarosa.LocalEntitiesExternalInstanceParserFactory
+import org.odk.collect.entities.javarosa.intance.LocalEntitiesExternalInstanceParserFactory
+import org.odk.collect.entities.javarosa.parse.EntityXFormParserFactory
 import org.odk.collect.metadata.PropertyManager
 import org.odk.collect.settings.SettingsProvider
 import org.odk.collect.settings.keys.ProjectKeys
@@ -40,7 +40,10 @@ class JavaRosaInitializer(
         )
 
         // Configure default parser factory
-        val entityXFormParserFactory = EntityXFormParserFactory(XFormParserFactory())
+        val entityXFormParserFactory =
+            EntityXFormParserFactory(
+                XFormParserFactory()
+            )
         val dynamicPreloadXFormParserFactory =
             DynamicPreloadXFormParserFactory(entityXFormParserFactory)
