+++ collect_app/src/main/java/org/odk/collect/android/javarosawrapper/FormController.kt
@@ -9,7 +9,7 @@ import org.javarosa.form.api.FormEntryCaption
 import org.javarosa.form.api.FormEntryPrompt
 import org.odk.collect.android.exception.JavaRosaException
 import org.odk.collect.android.formentry.audit.AuditEventLogger
-import org.odk.collect.entities.javarosa.EntitiesExtra
+import org.odk.collect.entities.javarosa.finalization.EntitiesExtra
 import java.io.File
 import java.io.IOException
