+++ entities/src/main/java/org/odk/collect/entities/LocalEntityUseCases.kt
@@ -3,8 +3,8 @@ package org.odk.collect.entities
 import org.javarosa.core.model.instance.CsvExternalInstance
 import org.javarosa.core.model.instance.TreeElement
 import org.odk.collect.entities.browser.EntityItemElement
-import org.odk.collect.entities.javarosa.EntitiesExtra
-import org.odk.collect.entities.javarosa.EntityAction
+import org.odk.collect.entities.javarosa.finalization.EntitiesExtra
+import org.odk.collect.entities.javarosa.spec.EntityAction
 import org.odk.collect.entities.storage.EntitiesRepository
 import org.odk.collect.entities.storage.Entity
 import java.io.File
