--- entities/src/main/java/org/odk/collect/entities/javarosa/EntitiesExtra.kt
@@ -1,3 +1,0 @@
-package org.odk.collect.entities.javarosa
-
-data class EntitiesExtra(val entities: List<Entity>)
