+++ entities/src/main/java/org/odk/collect/entities/javarosa/finalization/EntitiesExtra.kt
@@ -1,0 +1,3 @@
+package org.odk.collect.entities.javarosa.finalization
+
+data class EntitiesExtra(val entities: List<Entity>)
