+++ entities/src/main/java/org/odk/collect/entities/javarosa/spec/EntityAction.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.entities.javarosa
+package org.odk.collect.entities.javarosa.spec
 
 enum class EntityAction {
     CREATE,
