+++ entities/src/main/java/org/odk/collect/entities/javarosa/spec/UnrecognizedEntityVersionException.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.entities.javarosa
+package org.odk.collect.entities.javarosa.spec
 
 import org.javarosa.xform.parse.XFormParser
