+++ entities/src/test/java/org/odk/collect/entities/LocalEntityUseCasesTest.kt
@@ -7,8 +7,8 @@ import org.hamcrest.Matchers.containsInAnyOrder
 import org.hamcrest.Matchers.equalTo
 import org.junit.Test
 import org.odk.collect.entities.browser.EntityItemElement
-import org.odk.collect.entities.javarosa.EntitiesExtra
-import org.odk.collect.entities.javarosa.EntityAction
+import org.odk.collect.entities.javarosa.finalization.EntitiesExtra
+import org.odk.collect.entities.javarosa.spec.EntityAction
 import org.odk.collect.entities.storage.EntitiesRepository
 import org.odk.collect.entities.storage.Entity
 import org.odk.collect.entities.storage.Entity.State.ONLINE
@@ -23,7 +23,14 @@ class LocalEntityUseCasesTest {
     @Test
     fun `updateLocalEntitiesFromForm does not save updated entity that doesn't already exist`() {
         val entity =
-            org.odk.collect.entities.javarosa.Entity(EntityAction.UPDATE, "things", "1", "1", 1, emptyList())
+            org.odk.collect.entities.javarosa.finalization.Entity(
+                EntityAction.UPDATE,
+                "things",
+                "1",
+                "1",
+                1,
+                emptyList()
+            )
         val formEntities =
             EntitiesExtra(
                 listOf(entity)
@@ -37,7 +44,14 @@ class LocalEntityUseCasesTest {
     @Test
     fun `updateLocalEntitiesFromForm does not save entity that doesn't have an ID`() {
         val entity =
-            org.odk.collect.entities.javarosa.Entity(EntityAction.CREATE, "things", null, "1", 1, emptyList())
+            org.odk.collect.entities.javarosa.finalization.Entity(
+                EntityAction.CREATE,
+                "things",
+                null,
+                "1",
+                1,
+                emptyList()
+            )
         val formEntities =
             EntitiesExtra(
                 listOf(entity)
