+++ entities/src/test/java/org/odk/collect/entities/javarosa/EntitiesTest.java
@@ -26,6 +26,11 @@ import org.javarosa.xform.util.XFormUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.odk.collect.entities.javarosa.finalization.EntitiesExtra;
+import org.odk.collect.entities.javarosa.finalization.Entity;
+import org.odk.collect.entities.javarosa.finalization.EntityFormFinalizationProcessor;
+import org.odk.collect.entities.javarosa.parse.EntityXFormParserFactory;
+import org.odk.collect.entities.javarosa.spec.EntityAction;
 
 import java.io.IOException;
 import java.util.List;
