+++ entities/src/test/java/org/odk/collect/entities/javarosa/EntityFormParserTest.java
@@ -3,13 +3,15 @@ package org.odk.collect.entities.javarosa;
 import org.javarosa.core.model.data.IntegerData;
 import org.javarosa.core.model.instance.TreeElement;
 import org.junit.Test;
+import org.odk.collect.entities.javarosa.spec.EntityAction;
+import org.odk.collect.entities.javarosa.spec.EntityFormParser;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.odk.collect.entities.javarosa.EntityConstants.ATTRIBUTE_CREATE;
-import static org.odk.collect.entities.javarosa.EntityConstants.ATTRIBUTE_UPDATE;
-import static org.odk.collect.entities.javarosa.EntityConstants.ELEMENT_ENTITY;
-import static org.odk.collect.entities.javarosa.EntityConstants.ELEMENT_LABEL;
+import static org.odk.collect.entities.javarosa.spec.EntityConstants.ATTRIBUTE_CREATE;
+import static org.odk.collect.entities.javarosa.spec.EntityConstants.ATTRIBUTE_UPDATE;
+import static org.odk.collect.entities.javarosa.spec.EntityConstants.ELEMENT_ENTITY;
+import static org.odk.collect.entities.javarosa.spec.EntityConstants.ELEMENT_LABEL;
 
 public class EntityFormParserTest {
